==> 1_Data_Understanding.ipynb <==
# 1: Data Understanding

Di notebook ini cuma liat sekilas datanya, lalu menentukan ha

import pandas as pd
import numpy as np

df_train = pd.read_csv('./dataset/train.csv')
df_train['datetime_iso'] = pd.to_datetime(df_train['datetime_iso'])

df_test = pd.read_csv('./dataset/test.csv')
df_test['datetime_iso'] = pd.to_datetime(df_test['datetime_iso'])

df_train.head()

df_train.info()

# Kesimpulan:

Hasil lihat data sekilas:
- Datanya time series dengan feature lainnya
- Hampir semua tipedatanya masih bertipe object
- Berantakan, perlu dilakukan cleaning
- Karena time series dimulai dari tahun 1979, dipastikan datanya cuma gimik -> data sintesis
