==> 2_Data_Cleaning.ipynb <==
# 2. Data Cleaning

- Notebook ini berfokus merapikan data sehingga dapat diolah
- Notebook ini TIDAK melakukan preprocessing yang merubah informasi dari data


Random insight note:
- Ada temp yang sampe lebih dari 100 awokwokwok, antara outlier atau jgn jgn farenheit

## Kesimpulan setelah sampai akhir notebook: 
- Data numeris yang awalnya bertipe data objek, sudah jadi float
- Data yang formatnya belum sama, sudah disamakan
- Buat feature yang masih bertipe object, itu karena pada dasarnya = 0 lol, tapiii karena ini datanya sintesis, bisa aja masih ada korelasinya
- Data disimpan di train_cleaned.csv dan test_cleaned.csv

import pandas as pd
import numpy as np
import re

df_train = pd.read_csv('./dataset/train.csv')
df_train['datetime_iso'] = pd.to_datetime(df_train['datetime_iso'])

df_test = pd.read_csv('./dataset/test.csv')
df_test['datetime_iso'] = pd.to_datetime(df_test['datetime_iso'])

df_test.head(50)

## Cleaning Feature 'Temp'

Kita ekstrak angkanya dari string. 

Ini fungsinya bisa dipake lagi

# Function to extract numeric values and convert to float
def clean_and_convert_temperature(temperature_str):
    if temperature_str is np.nan:
        return None
    
    # Use regular expressions to extract numeric values
    temperature_str = temperature_str.replace(',', '.')  # Replace comma with period (if present)
    numeric_part = re.search(r'[\d.]+', temperature_str)
    
    if numeric_part:
        numeric_value = float(numeric_part.group())
        return numeric_value
    else:
        return None

# Apply the cleaning and conversion function to the column
df_train['temp'] = df_train['temp'].apply(clean_and_convert_temperature)
df_test['temp'] = df_test['temp'].apply(clean_and_convert_temperature)

## Cleaning Feature 'Visibility'

df_train.visibility.value_counts()

Ini maksudnya gimana bjir, skip dulu deh

## Cleaning Feature 'd_point, feels, min_temp, max_temp'

Ini mirip sama yg temp jd sekalian aja pake fungsi dari temp

# Cleaning d_point
df_train['d_point'] = df_train['d_point'].apply(clean_and_convert_temperature)
df_test['d_point'] = df_test['d_point'].apply(clean_and_convert_temperature)

# Cleaning feels
df_train['feels'] = df_train['feels'].apply(clean_and_convert_temperature)
df_test['feels'] = df_test['feels'].apply(clean_and_convert_temperature)

# Cleaning min_temp
df_train['min_temp'] = df_train['min_temp'].apply(clean_and_convert_temperature)
df_test['min_temp'] = df_test['min_temp'].apply(clean_and_convert_temperature)

# Cleaning max_temp
df_train['max_temp'] = df_train['max_temp'].apply(clean_and_convert_temperature)
df_test['max_temp'] = df_test['max_temp'].apply(clean_and_convert_temperature)

## Cleaning Feature 'prssr'

Intinya diambil angkanya, ternyata simple bjir

df_train['prssr'] = df_train['prssr'].str.extract(r'(\d+\.\d+|\d+)').astype(float)
df_test['prssr'] = df_test['prssr'].str.extract(r'(\d+\.\d+|\d+)').astype(float)

## Cleaning feature 'sea_level' and 'ground_level'

df_train.sea_level.value_counts()

df_test.sea_level.value_counts()

df_train.grnd_level.value_counts()

df_test.grnd_level.value_counts()

Intinya datanya gaada ga sih ?! // kena preng

## Cleaning feature 'hum'

Disini datanya ada yg pake persen (%), jd mari lah

df_train['hum'] = df_train['hum'].str.extract(r'(\d+\.\d+|\d+)').astype(float)
df_test['hum'] = df_test['hum'].str.extract(r'(\d+\.\d+|\d+)').astype(float)

## Cleaning feature 'wind_spd'

df_train['wind_spd'] = df_train['wind_spd'].str.extract(r'(\d+\.\d+)').astype(float)
df_test['wind_spd'] = df_test['wind_spd'].str.extract(r'(\d+\.\d+)').astype(float)

## Cleaning feature 'wind_deg'

df_train['wind_deg'] = df_train['wind_deg'].str.replace('[^0-9.]', '', regex=True).astype(float)
df_test['wind_deg'] = df_test['wind_deg'].str.replace('[^0-9.]', '', regex=True).astype(float)

## Cleaning label 'rain_1h'

# Replace 'zero' with '0'
df_train['rain_1h'] = df_train['rain_1h'].replace('zero', '0')

# Replace all whitespace characters with '0'
df_train['rain_1h'] = df_train['rain_1h'].str.replace(r'\s+', '0', regex=True)

# Replace all to numeric
df_train['rain_1h'] = df_train['rain_1h'].str.extract(r'(-?\d+\.\d+|\d+)').astype(float)


### Cleaning feature 'rain_3h', 'snow_1h', 'snow_3h'

df_train.rain_3h.value_counts()

df_test.rain_3h.value_counts()

df_train.snow_1h.value_counts()

df_test.snow_1h.value_counts()

df_train.snow_3h.value_counts()

df_test.snow_3h.value_counts()

Homto gaje lagi

## Cleaning feature 'clouds'

df_train['clouds'] = df_train['clouds'].str.replace('[^0-9.]', '', regex=True).astype(float)
df_test['clouds'] = df_test['clouds'].str.replace('[^0-9.]', '', regex=True).astype(float)

## Cek terakhir

df_train.info()

df_test.info()

Data siap dipake nih bang

## Save ke csv

df_train.to_csv('./dataset/train_cleaned.csv', index=False)
df_test.to_csv('./dataset/test_cleaned.csv', index=False)


# Kesimpulan:

- Data numeris yang awalnya bertipe data objek, sudah jadi float
- Data yang formatnya belum sama, sudah disamakan
- Buat feature yang masih bertipe object, itu karena pada dasarnya = 0 lol, tapiii karena ini datanya sintesis, bisa aja masih ada korelasinya
- Data disimpan di train_cleaned.csv dan test_cleaned.csv
