==> 3_Model_Treshold.ipynb <==
### Kita Coba bikin model pake pycaret full auto

Kesimpulan:
- untuk sementara feature feature gaje dinyatakan sebagai sampah

import pandas as pd

df_train = pd.read_csv('./dataset/train_cleaned.csv')
df_train['datetime_iso'] = pd.to_datetime(df_train['datetime_iso'])

df_test = pd.read_csv('./dataset/test_cleaned.csv')
df_test['datetime_iso'] = pd.to_datetime(df_test['datetime_iso'])

df_train.head()

## Percobaan 1: Drop datetime only

df_train = df_train.drop(['datetime','datetime_iso','time-zone'],axis = 1)
df_test = df_test.drop(['datetime','datetime_iso','time-zone'],axis = 1)

%pip install pycaret

from pycaret.regression import *
s = setup(df_train, target = 'rain_1h', session_id = 123)

best = compare_models()

## Percobaan 2: Drop feature gajelas

df_train = df_train.drop(['visibility','sea_level','grnd_level','rain_3h','snow_1h','snow_3h'],axis = 1)
df_test = df_test.drop(['visibility','sea_level','grnd_level','rain_3h','snow_1h','snow_3h'],axis = 1)

s2 = setup(df_train, target = 'rain_1h', session_id = 123)

best2 = compare_models()
