==> 5_0_Data_Preprocessing.ipynb <==
# 5. Data Preprocessing

Disini kita bakal:
- Null Handling
- Outlier Handling

import pandas as pd
import numpy as np

df_train = pd.read_csv('./dataset/train_cleaned.csv')
df_train['datetime_iso'] = pd.to_datetime(df_train['datetime_iso'])

df_test = pd.read_csv('./dataset/test_cleaned.csv')
df_test['datetime_iso'] = pd.to_datetime(df_test['datetime_iso'])

df_train.info()

## Drop unnecesary Feature

df_train = df_train.drop(['datetime','time-zone','visibility','sea_level','grnd_level','rain_3h','snow_1h','snow_3h'],axis = 1)
df_test = df_test.drop(['datetime','time-zone','visibility','sea_level','grnd_level','rain_3h','snow_1h','snow_3h'],axis = 1)

df_train.head()

# Null Handling

## Null Handling: Median

Hanya ada satu feature yang memiliki null, yaitu wind_spd.

Oh ternyata ada d_point yang null di df_test

df_train['wind_spd'] = df_train['wind_spd'].interpolate(method='linear')
df_test['wind_spd'] = df_test['wind_spd'].interpolate(method='linear')
df_test['d_point'] = df_test['d_point'].interpolate(method='linear')

df_train.info()

df_test.info()

# Outlier Handling

## Outlier Handling: Interpolation

Konsep Kerja: Ada outlier -> Dijadiin Null -> Diinterpolasi lagi hehe

#Outlier Detection and Replacement (Quantile and Median Method)

df_columns_outlier_handle = ['temp',
 'd_point',
 'feels',
 'min_temp',
 'max_temp',
 'prssr',
 'hum',
 'wind_spd',
 'wind_deg',
 'clouds']

# df_outhand = df_train.copy()

for columnName in df_columns_outlier_handle:
  #detecting outliers (1.5 below/above)
  Q1 = df_train[columnName].quantile(0.25)
  Q3 = df_train[columnName].quantile(0.75)
  # print(Q1, Q3)
  IQR = Q3-Q1
  # print(IQR)
  lwr_bound = Q1-(1.5*IQR)
  upr_bound = Q3+(1.5*IQR)

  # #replace with median 
  median = np.median(df_train[columnName])

  df_train.loc[(df_train[columnName] < lwr_bound) | (df_train[columnName] > upr_bound), columnName] = np.nan
    
for columnName in df_columns_outlier_handle:
  #detecting outliers (1.5 below/above)
  Q1 = df_test[columnName].quantile(0.25)
  Q3 = df_test[columnName].quantile(0.75)
  # print(Q1, Q3)
  IQR = Q3-Q1
  # print(IQR)
  lwr_bound = Q1-(1.5*IQR)
  upr_bound = Q3+(1.5*IQR)

  # #replace with median 
  median = np.median(df_test[columnName])

  df_test.loc[(df_test[columnName] < lwr_bound) | (df_test[columnName] > upr_bound), columnName] = np.nan

f_train.info()

df_test.info()

### Isi elemen yang jadi null dengan interpolasi

for columnName in df_columns_outlier_handle:
    df_train[columnName] = df_train[columnName].interpolate(method='linear')
    df_test[columnName] = df_test[columnName].interpolate(method='linear')

df_train.info()

df_test.info()

### Save in csv

df_train.to_csv('./dataset/train_preprocessed.csv', index=False)
df_test.to_csv('./dataset/test_preprocessed.csv', index=False)
