==> 5_1_Handle_Missing_Values.ipynb <==
## Di sini masih pake cara manual krn sy ga tau lol aph itu interpolasi

import pandas as pd

df_train_nan = pd.read_csv('./dataset/train_nan.csv')
df_train_nan['datetime_iso'] = pd.to_datetime(df_train_nan['datetime_iso'])

df_train_median = pd.read_csv('./dataset/train_median.csv')
df_train_median['datetime_iso'] = pd.to_datetime(df_train_median['datetime_iso'])

df_train_nan = df_train_nan.drop(['visibility','sea_level','grnd_level','rain_3h','snow_1h','snow_3h'],axis = 1)
df_train_median = df_train_median.drop(['visibility','sea_level','grnd_level','rain_3h','snow_1h','snow_3h'],axis = 1)


null_cols = df_train_nan.columns[df_train_nan.isna().any()].tolist()
null_cols_2 = df_train_median.columns[df_train_nan.isna().any()].tolist()

df_train_nan.isna().sum()

df_train_median.isna().sum()

## Fungsi di sini

df_train_nan['date'] = df_train_nan['datetime_iso'].dt.date
df_train_median['date'] = df_train_median['datetime_iso'].dt.date

def filling_missing_values(df, col: str):
  # manual filling
  df[col] = df[col].fillna(df[col].shift() + df[col].shift(-1) / 2)  
  group_by_date = round(df.groupby('date')['wind_spd'].mean(), 2)
  df[col] = df[col].fillna(df['date'].map(group_by_date))

def filling_missing_values_with_interpolation(df):
  # time interpolation, nek pake ini yang bawah di-comment
  df = df.set_index('datetime_iso') 
  df = df.interpolate(method='time')

  # selain itu
  df = df.interpolate(method='backfill')


df_list = [df_train_nan, df_train_median]

for col in null_cols:
  for item in df_list:
    # ganti di sini adik2
    filling_missing_values(item, col)
    print(item.isna().sum())
  


# drop helper column
df_train_nan = df_train_nan.drop('date', axis=1)
df_train_median = df_train_median.drop('date', axis=1)

df_train_nan.to_csv('./dataset/train_nan_cleaned.csv', index=False)
df_train_median.to_csv('./dataset/train_median_cleaned.csv', index=False)

## Di bawah ini ga kepake xD

index = df_train_nan[df_train_nan['wind_spd'].isna()].index

# liat data yang missing sama intervalnya
missing_with_intervals = df_train_nan[(df_train_nan.index.isin(index)) | (df_train_nan.index.isin(index-1)) | (df_train_nan.index.isin(index+1))]
missing_with_intervals.head(12)


# isi missing value dengan nilai rata-rata dari data sebelum dan sesudahnya
df_train_nan['wind_spd'] = df_train_nan['wind_spd'].fillna((df_train_nan['wind_spd'].shift() + df_train_nan['wind_spd'].shift(-1))/2)

# cek lagi missing value
df_train_nan.iloc[15:22]

df_train_nan.isna().sum()

masih ada karena ya bisa aja samping kanan kiri belio isinya NaN :D

index_layer2 = df_train_nan[df_train_nan['wind_spd'].isna()].index

# ambil satu ajah ckup
start_interval = index_layer2[0] - 4
end_interval = index_layer2[0] + 4

df_train_nan.iloc[start_interval:end_interval]

betyul sebelahnya NaN, sbenere bisa aja ambil kirinya tapi kalo NaN lagi gimana? ambil kirinya kagi? capek ga luwh

df_train_nan['date'] = df_train_nan['datetime_iso'].dt.date

# jd gweh ngegroup by date then cari mean-nyah
group_by_date = round(df_train_nan.groupby('date')['wind_spd'].mean(), 2)

group_by_date.head(5)

# isi missing value dengan rata2 di dalam group_by_date
df_train_nan['wind_spd'] = df_train_nan['wind_spd'].fillna(df_train_nan['date'].map(group_by_date))

df_train_nan.iloc[start_interval:end_interval]
