==> 6_2_Feature_Extraction_Lag.ipynb <==
# 6.2. Create Lag Feature

Point of this notebook: Mana yang lebih baik, membuat feature lag dengan data 6 tahun lalu tetapi mengorbankan data row selama 6 tahun, atau apa adanya?

### Create Feature: Lag 6 years ago

import pandas as pd

df_train = pd.read_csv('./dataset/train_ts_feature_added.csv')
df_train['datetime_iso'] = pd.to_datetime(df_train['datetime_iso'])
df_train['source'] = 'df_train'

df_test = pd.read_csv('./dataset/test_ts_feature_added.csv')
df_test['datetime_iso'] = pd.to_datetime(df_test['datetime_iso'])
df_test['source'] = 'df_test'

df_train.info()

df_test.info()

df = pd.concat([df_train, df_test], axis=0)  # Stacking DataFrames vertically

df.info()

# Specify the number of years to lag
lag_years = 6

# Calculate the lagged date
df['lag_rain_1h_6_years_date'] = df['datetime_iso'] - pd.DateOffset(years=lag_years)

df['lag_rain_1h_6_years_date'] = pd.to_datetime(df['lag_rain_1h_6_years_date'])

# Merge the DataFrame with itself to get lag_rain_1h_6_years_value
df =df.merge(df[['datetime_iso', 'rain_1h']], left_on='lag_rain_1h_6_years_date', right_on='datetime_iso', suffixes=('', '_lag'))

# Rename the 'rain_1h_lag' column to 'lag_rain_1h_6_years_value'
df.rename(columns={'rain_1h_lag': 'lag_rain_1h_6_years_value'}, inplace=True)

# Drop the duplicated datetime_iso column
df.drop(columns='datetime_iso_lag', inplace=True)

df.info()

df_train = df[df['source'] == 'df_train']
df_test = df[df['source'] == 'df_test']

# Remove the 'source' column if you no longer need it
df_train = df_train.drop(columns=['source'])
df_test = df_test.drop(columns=['source','rain_1h'])

df_train.sample(30)

df_train.info()

df_test.head()

df_test.info()

df_train.to_csv('./dataset/train_ts_feature_added_2.csv', index=False)
df_test.to_csv('./dataset/test_ts_feature_added_2.csv', index=False)
