==> 8_1_Modeling_ML.ipynb <==
import pandas as pd

df_train = pd.read_csv('./dataset/train_ts_feature_added.csv')
df_train['datetime_iso'] = pd.to_datetime(df_train['datetime_iso'])

df_test = pd.read_csv('./dataset/test_ts_feature_added.csv')
df_test['datetime_iso'] = pd.to_datetime(df_test['datetime_iso'])
df_test_date = df_test['datetime_iso']

df_train.head()

## Modelling using Pycaret:Catboost

df_test.info()

df_train = df_train.drop(['datetime_iso'],axis = 1)
df_test = df_test.drop(['datetime_iso','rain_1h'],axis = 1)

from pycaret.regression import *
s = setup(df_train, target = 'rain_1h', session_id = 41)

#best = compare_models()

catboost = create_model('catboost')

predictions = predict_model(catboost, data=df_test)
predictions.head()

submission = pd.DataFrame()
submission['datetime_iso'] = df_test_date
submission['rain_1h'] = predictions['prediction_label']
submission.head()

submission.info()

submission.to_csv('./submission/submission_7.csv', index=False)
